Move server handlers into a server subdir
