+++ go_api/handler/server/favoriteHandler.go
@@ -1,4 +1,4 @@
-package handler
+package server
 
 import (
 	"errors"
