+++ go_api/handler/server/userHandler.go
@@ -1,4 +1,4 @@
-package handler
+package server
 
 import (
 	"errors"
